点击去结算，到达订单核对页面，同时生成校验码。点击提交订单，（后台为防止同一订单多次提交进行校验码检查，查完立即销毁保证页面只能被提交一次）。只需要传入收货信息id，后台查询购物车中选中的商品来生成订单，订单详情，（生成订单过程中进行订单详情的价格（与sku的price进行对比），库存校验，只要有一个订单详情校验失败，直接返回核对订单页，让用户重新提交），订单生成后删除购物车内容，调用cartService，根据MemberId，SkuId删除。

+++ gmall-api/src/main/java/com/atguigu/gmall/bean/OmsCartItem.java
@@ -15,7 +15,7 @@ public class OmsCartItem implements Serializable{
     private String productSkuId;
     private String memberId;
     @Column
-    private Integer quantity;
+    private BigDecimal quantity;
     private BigDecimal price;
     private String sp1;
     private String sp2;
@@ -76,11 +76,11 @@ public class OmsCartItem implements Serializable{
         this.memberId = memberId;
     }
 
-    public Integer getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(BigDecimal quantity) {
         this.quantity = quantity;
     }
